% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decontaminate.R
\name{decontaminate}
\alias{decontaminate}
\title{Simulambient DecontX Wrapper}
\usage{
decontaminate(dataset, contamination = NULL, cellTypes = NULL, batch = NULL)
}
\arguments{
\item{dataset}{An n x k matrix-like object where n = number of genes and k = number of nuclei to be decontaminated.}

\item{contamination}{An n x j matrix-like object where n = number of genes and j = number of ambient mRNA droplets. The droplets in this matrix are believed to be by the user to be composed of ambient mRNA. Including droplets beneath a specific UMI cutoff (i.e. 100) is a basic idea to get ambient mRNA droplets.}

\item{cellTypes}{A vector of length k correpsonding to the cell types of the nuclei in \code{dataset}.}

\item{batch}{A vector of length k corresponding to batch information of the nuclei in \code{dataset}.}

\item{...}{Any extra parameters passed to DecontX.}
}
\value{
A sparse matrix of class dgCMatrix with the decontaminated counts for the experiment.
}
\description{
decontaminate will perform snRNA-seq decontamination on an snRNA-seq dataset using a contamination reference and cell types.
}
