% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_params.R
\name{estimate_params}
\alias{estimate_params}
\title{Estimate Parameters from Dataset}
\usage{
estimate_params(dataset, metadata = NULL, metadata_indices = NULL)
}
\arguments{
\item{dataset}{An n x k matrix-like object where n = number of genes and k = number of nuclei. Assumed to be demultiplexed and cell type-annotated.}

\item{metadata}{A matrix containing metadata information provided by the user. Has k rows.}

\item{metadata_indices}{A vector informing which nuclei belong to which metadata group.}
}
\value{
A list containing two values: a matrix \code{geneMeans} with mean gene expression for each gene for each metadata group, and a matrix \code{geneSDs} with the standard deviation gene expression for each gene in each metadata group.
}
\description{
estimate_params will take a snRNA-seq dataset as input and estimate parameters for downstream analysis.
}
