% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_umis.R
\name{get_umis}
\alias{get_umis}
\title{Calculate Pre-DecontX Counts}
\usage{
get_umis(dataset, metadata = NULL, cellTypes = NULL, sex = NULL)
}
\arguments{
\item{dataset}{An n x k matrix-like object where n = number of genes and k = number of nuclei. Assumed to be demultiplexed and cell type-annotated.}

\item{metadata}{A vector of strings corresponding to the metadata made available by the user. This argument will be handled inside the main function \code{simulambient} and is the return value of \code{check_args}.}

\item{cellTypes}{A vector of length k that provides the cell type of each nucleus in \code{dataset}. If all nuclei are from one type or cell type is not to be accounted for, parameter can be left blank.}

\item{sex}{A vector of length k that provides the associated sex of each droplet in \code{dataset}.}
}
\value{
A list containing four values. The first value is a matrix \code{geneMeans} that lists the mean counts value for each gene (for that particular metadata, if applicable). The second value is a matrix \code{geneMeans} that lists the standad deviation for each gene (for that particular metadata, if applicable). The third value is a matrix \code{umis} that lists the UMI counts for each droplet (for that particular metadata, if applicable). The fourth value is a matrix-like object that holds the metadata correpsonding to the columns/rows of the other returned objects.
}
\description{
get_umis will get the pre-DecontX counts for each nucleus so that the scaling factor for Simulambient can be computed.
}
