% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulambient.R
\name{simulambient}
\alias{simulambient}
\title{Simulambient}
\usage{
simulambient(
  dataset,
  contamination_levels = 1,
  contamination = NULL,
  metadata = NULL,
  cellTypes = NULL,
  batch = NULL,
  bgBatch = NULL,
  seed = 615,
  ...
)
}
\arguments{
\item{dataset}{An n x k matrix-like object where n = number of genes and k = number of nuclei to serve as the basis of the simulation.}

\item{contamination_levels}{A numeric value or a vector of numeric values indicating the desired levels of contamination that the user would like to simulate. A value of 1 indicates the user wants to simulate the level of contamination detected in the input dataset; a value of 0.5 indicates that the user wants to simulate half of the contamination detected in the original dataset, etc.}

\item{contamination}{An n x j matrix-like object where n = number of genes and j = number of ambient mRNA droplets. The droplets in this matrix are believed to be by the user to be composed of ambient mRNA. Including droplets beneath a specific UMI cutoff (i.e. 100) is a basic idea to get ambient mRNA droplets.}

\item{metadata}{A matrix with k rows that corresponds to metadata of the k nuclei provided in \code{dataset}. Please see \code{cellTypes}, \code{batch}, and \code{bgBatch}. Each column should represent a different aspect of metadata, i.e. sex. It should be noted that the user should be rather parsimonious when choosing which metadata to include. Including too many parameters will increase the number of metadata permutations, which will increase runtime and potentially "overfit" the simulations. Cell types is strongly recommended, anything else is up to the user's discretion.}

\item{cellTypes}{A vector of length k should cell type information not be provided in \code{metadata}. If it is provided, this parameter should be a string of the column name that represents the cell type information in \code{metadata}. If cell type information is not desired to be included, default is NULL.}

\item{batch}{A vector of length k that provides the batch of the nucleus in \code{dataset}. If there is no batch effect, parameter can be left blank. This parameter must be provided if \code{bgBatch} is provided (as required by DecontX). Please do not include batch information in the \code{metadata} parameter.}

\item{bgBatch}{A vector of length j corresponding to batch information of the ambient mRNA droplets in \code{contamination}. This parameter must be provided if \code{batch} is provided (as required by DecontX).}

\item{seed}{Seed for random number generation. Can be set by the user. Defaults to 615.}

\item{...}{Any extra parameters passed to DecontX from the \code{celda} package.}
}
\description{
simulambient is the main function of the package. It calls the helper functions to simulate snRNA-seq counts with varying levels of contamination.
}
