% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_counts.R
\name{simulate_counts}
\alias{simulate_counts}
\title{Simulate snRNA-seq Counts with Negative Binomial}
\usage{
simulate_counts(geneMeans, geneSDs, numNucleiPerGroup, seed)
}
\arguments{
\item{geneMeans}{Matrix-like object of gene expression means for each metadata group. (Rows = genes, columns = metadata groups)}

\item{numNucleiPerGroup}{Vector of integers correpsonding to how many nuclei need to be simulated within each metadata group.}

\item{seed}{Seed for random number generation. Can be set by the user. Defaults to 615.}
}
\description{
simulate_counts simulates snRNA-seq counts given parameters for the negative binomial.
}
