% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_args.R
\name{check_args}
\alias{check_args}
\title{Check User Arguments}
\usage{
check_args(cellTypes = NULL, batch = NULL, sample = NULL)
}
\arguments{
\item{cellTypes}{A vector of length k that provides the cell type of each nucleus in \code{dataset}. If all nuclei are from one type or cell type is not to be accounted for, parameter can be left blank.}

\item{batch}{A vector of length k that provides the batch of the nucleus in \code{dataset}. If there is no batch effect, parameter can be left blank.}

\item{sample}{A vector of length k that provides the associated sex of each droplet in \code{dataset}.}
}
\value{
A list of strings containing the valid metadata provided by the user.
}
\description{
check_args will take any cellTypes, batch, and sample arguments provided by the user and check them for validity.
}
