% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_args.R
\name{check_args}
\alias{check_args}
\title{Check User Arguments}
\usage{
check_args(
  dataset,
  contamination = NULL,
  cellTypes = NULL,
  batch = NULL,
  bgBatch = NULL,
  sex = NULL
)
}
\arguments{
\item{dataset}{An n x k matrix-like object where n = number of genes and k = number of nuclei to be decontaminated.}

\item{contamination}{An n x j matrix-like object where n = number of genes and j = number of ambient mRNA droplets. The droplets in this matrix are believed to be by the user to be composed of ambient mRNA. Including droplets beneath a specific UMI cutoff (i.e. 100) is a basic idea to get ambient mRNA droplets.}

\item{cellTypes}{A vector of length k that provides the cell type of each nucleus in \code{dataset}. If all nuclei are from one type or cell type is not to be accounted for, parameter can be left blank.}

\item{batch}{A vector of length k that provides the batch of the nucleus in \code{dataset}. If there is no batch effect, parameter can be left blank.}

\item{bgBatch}{A vector of length j corresponding to batch information of the ambient mRNA droplets in \code{contamination}.}

\item{sex}{A vector of length k that provides the associated sex of each droplet in \code{dataset}.}
}
\value{
A list of strings containing the valid metadata provided by the user.
}
\description{
check_args will take any cellTypes, batch, and sex arguments provided by the user and check them for validity.
}
