% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_ambient_params.R
\name{estimate_ambient_params}
\alias{estimate_ambient_params}
\title{Estimate Ambient Parameters from Contamination}
\usage{
estimate_ambient_params(contamination, bgBatch = NULL)
}
\arguments{
\item{contamination}{An n x j matrix-like object where n = number of genes and j = number of ambient mRNA droplets. The droplets in this matrix are believed to be by the user to be composed of ambient mRNA. Including droplets beneath a specific UMI cutoff (i.e. 100) is a basic idea to get ambient mRNA droplets.}

\item{bgBatch}{A vector of length j corresponding to batch information of the ambient mRNA droplets in \code{contamination}.}
}
\value{
A list containing two values. The first value is a matrix \code{geneMeans} that lists the mean counts value for each gene (for that particular metadata, if applicable). The second value is a matrix \code{geneMeans} that lists the standad deviation for each gene (for that particular metadata, if applicable).
}
\description{
estimate_ambient_params will take a snRNA-seq contamination dataset as input and estimate parameters for downstream analysis.
}
