% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_params.R
\name{estimate_params}
\alias{estimate_params}
\title{Estimate Parameters from Dataset}
\usage{
estimate_params(
  dataset,
  metadata = NULL,
  cellTypes = NULL,
  batch = NULL,
  sample = NULL
)
}
\arguments{
\item{dataset}{An n x k matrix-like object where n = number of genes and k = number of nuclei. Assumed to be demultiplexed and cell type-annotated.}

\item{metadata}{A vector of strings corresponding to the metadata made available by the user. This argument will be handled inside the main function \code{simulambient} and is the return value of \code{check_args}.}

\item{cellTypes}{A vector of length k that provides the cell type of each nucleus in \code{dataset}. If all nuclei are from one type or cell type is not to be accounted for, parameter can be left blank.}

\item{batch}{A vector of length k that provides the batch of the nucleus in \code{dataset}. If there is no batch effect, parameter can be left blank.}

\item{sample}{A vector of length k that provides the associated sex of each droplet in \code{dataset}.}
}
\value{
A list containing two values. The first value is a matrix \code{geneMeans} that lists the mean counts value for each gene (for that particular metadata, if applicable). The second value is a matrix \code{umis} that lists the UMI counts for each droplet (for that particular metadata, if applicable).
}
\description{
estimate_params will take a snRNA-seq dataset as input and estimate parameters for downstream analysis.
}
