% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulambient.R
\name{simulambient}
\alias{simulambient}
\title{Simulambient}
\usage{
simulambient(
  dataset,
  contamination_levels = 1,
  contamination = NULL,
  cellTypes = NULL,
  batch = NULL,
  bgBatch = NULL,
  sex = NULL,
  seed = 615,
  ...
)
}
\arguments{
\item{dataset}{An n x k matrix-like object where n = number of genes and k = number of nuclei to serve as the basis of the simulation.}

\item{contamination_levels}{A numeric value or a vector of numeric values indicating the desired levels of contamination that the user would like to simulate. A value of 1 indicates the user wants to simulate the level of contamination detected in the input dataset; a value of 0.5 indicates that the user wants to simulate half of the contamination detected in the original dataset, etc.}

\item{contamination}{An n x j matrix-like object where n = number of genes and j = number of ambient mRNA droplets. The droplets in this matrix are believed to be by the user to be composed of ambient mRNA. Including droplets beneath a specific UMI cutoff (i.e. 100) is a basic idea to get ambient mRNA droplets.}

\item{cellTypes}{A vector of length k that provides the cell type of each nucleus in \code{dataset}. If all nuclei are from one type or cell type is not to be accounted for, parameter can be left blank.}

\item{batch}{A vector of length k that provides the batch of the nucleus in \code{dataset}. If there is no batch effect, parameter can be left blank. This parameter must be provided if \code{bgBatch} is provided (as required by DecontX).}

\item{bgBatch}{A vector of length j corresponding to batch information of the ambient mRNA droplets in \code{contamination}. This parameter must be provided if \code{batch} is provided (as required by DecontX).}

\item{sex}{A vector of length k that provides the associated sex of each droplet in \code{dataset}.}

\item{seed}{Seed for random number generation. Can be set by the user. Defaults to 615.}

\item{...}{Any extra parameters passed to DecontX from the \code{celda} package.}
}
\description{
simulambient is the main function of the package. It calls the helper functions to simulate snRNA-seq counts with varying levels of contamination.
}
